backend/requirements.txt
pip==20.2.3
Django==3.1.3
psycopg2>=2.7,<3.0
pylint==2.5.0
pillow==7.1.2
djangorestframework==3.12
mysqlclient==2.0.1
soft-delete==0.2.2
django-oauth-toolkit==1.3.3
django-cors-headers==3.7.0
django-nested-admin==3.3.3
drf-yasg==1.20.0